Fix weather package name.
