+++ weather/src/main/java/com/group5/weather/providers/OpenWeatherApiWeatherResponseItem.kt
@@ -1,4 +1,4 @@
-package com.norfoe.weather.providers
+package com.group5.weather.providers
 
 import com.squareup.moshi.Json
