--- weather/src/main/java/com/norfoe/weather/Provider.kt
@@ -1,13 +1,0 @@
-package com.norfoe.weather
-
-interface Provider {
-    fun init();
-
-    /**
-     * Config is used to pass various data to the provider, e.g. an API key
-     */
-    fun setConfig(config: HashMap<String, String>)
-
-    suspend fun fetch(cityName: String? = ""): WeatherInfo
-    suspend fun fetch(lat: String? = "0", lon: String? = "0"): WeatherInfo
-}
